+++ miniver/app.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 # This file is part of 'miniver': https://github.com/jbweston/miniver
 
 import sys
@@ -38,7 +37,9 @@ _zipfile_root = "miniver-master"  # tied to the fact that we fetch master.zip
 # File templates
 _setup_template = textwrap.dedent(
     '''
-    def get_version_and_cmdclass(package_path):
+    from setuptools import setup
+
+    def get_version_and_cmdclass(pkg_path):
         """Load version.py module without importing the whole package.
 
         Template code from miniver
@@ -46,13 +47,13 @@ _setup_template = textwrap.dedent(
         import os
         from importlib.util import module_from_spec, spec_from_file_location
 
-        spec = spec_from_file_location("version", os.path.join(package_path, "_version.py"))
+        spec = spec_from_file_location("version", os.path.join(pkg_path, "_version.py"))
         module = module_from_spec(spec)
         spec.loader.exec_module(module)
-        return module.__version__, module.cmdclass
+        return module.__version__, module.get_cmdclass(pkg_path)
 
 
-    version, cmdclass = get_version_and_cmdclass("{package_dir}")
+    version, cmdclass = get_version_and_cmdclass(r"{package_dir}")
 
 
     setup(
@@ -201,11 +202,13 @@ def install(args):
             "You still have to copy the following snippet into your 'setup.py':"
         )
     )
-    print("\n".join((msg, _setup_template)).format(package_dir=package_dir))
+    # Use stdout for setup template only, so it can be redirected to 'setup.py'
+    print(msg.format(package_dir=package_dir), file=sys.stderr)
+    print(_setup_template.format(package_dir=package_dir), file=sys.stdout)
 
 
 def ver(args):
-    search_path = args.search_path
+    search_path = os.path.realpath(args.search_path)
     try:
         version_location, = glob.glob(
             os.path.join(search_path, "**", "_version.py"),
@@ -244,5 +247,7 @@ def main():
         args.dispatch(args)
 
 
+# This is needed when using the script directly from GitHub, but not if
+# miniver is installed.
 if __name__ == "__main__":
     main()